Only run blueprint add/remove operations in update when strictly necessary

+++ osu.Game/Screens/Edit/Compose/Components/BlueprintContainer.cs
@@ -10,6 +10,7 @@ using System.Linq;
 using JetBrains.Annotations;
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
+using osu.Framework.Caching;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Input;
@@ -243,11 +244,16 @@ namespace osu.Game.Screens.Edit.Compose.Components
                 UpdateSelectionFromDragBox(selectionBeforeDrag);
             }
 
-            SelectionBlueprints.AddRange(blueprintsToAdd);
-            blueprintsToAdd.Clear();
+            if (!blueprintsPending.IsValid)
+            {
+                SelectionBlueprints.AddRange(blueprintsToAdd);
+                blueprintsToAdd.Clear();
 
-            SelectionBlueprints.RemoveRange(blueprintsToRemove, true);
-            blueprintsToRemove.Clear();
+                SelectionBlueprints.RemoveRange(blueprintsToRemove, true);
+                blueprintsToRemove.Clear();
+
+                blueprintsPending.Validate();
+            }
         }
 
         /// <summary>
@@ -317,6 +323,7 @@ namespace osu.Game.Screens.Edit.Compose.Components
 
         #region Blueprint Addition/Removal
 
+        private readonly Cached blueprintsPending = new Cached();
         private readonly List<SelectionBlueprint<T>> blueprintsToAdd = new List<SelectionBlueprint<T>>();
         private readonly List<SelectionBlueprint<T>> blueprintsToRemove = new List<SelectionBlueprint<T>>();
 
@@ -335,6 +342,7 @@ namespace osu.Game.Screens.Edit.Compose.Components
             blueprint.Deselected += OnBlueprintDeselected;
 
             blueprintsToAdd.Add(blueprint);
+            blueprintsPending.Invalidate();
 
             if (SelectionHandler.SelectedItems.Contains(item))
                 blueprint.Select();
@@ -352,6 +360,7 @@ namespace osu.Game.Screens.Edit.Compose.Components
             blueprintToRemove.Deselected -= OnBlueprintDeselected;
 
             blueprintsToRemove.Add(blueprintToRemove);
+            blueprintsPending.Invalidate();
 
             if (movementBlueprints?.Any(m => m.blueprint == blueprintToRemove) == true)
                 finishSelectionMovement();
